api_service/src/chatbot/serializers/history.py
from rest_framework import serializers

from chatbot.models import History


class HistorySerializer(serializers.ModelSerializer):

    class Meta:
        model = History
        fields = ['id', 'query', 'answer', 'metadata']
