api_service/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.1.3
billiard==4.2.0
blis==0.7.11
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
cryptography==42.0.7
cymem==2.0.8
dataclasses-json==0.6.6
Deprecated==1.2.14
Django==4.2.11
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
dnspython==2.6.1
docx2txt==0.8
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.1
Faker==25.3.0
fastapi==0.110.3
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
graphlib_backport==1.1.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jwcrypto==1.5.6
kombu==5.3.7
kubernetes==29.0.0
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.60
language_data==1.2.0
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.18.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pillow==10.3.0
pip-chill==1.0.3
posthog==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg==3.1.18
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pysqlite-binary==0.5.1.3380300
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rank-bm25==0.2.2
redis==5.0.4
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
srsly==2.4.8
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.2
